renames classes and fragments
changes the default naming of the fragments to profile, map and battles
also changes the names and color of the bottom nav bar
