--- app/src/main/java/com/walkly/walkly/ui/home/HomeFragment.kt
@@ -1,267 +1,0 @@
-package com.walkly.walkly.ui.home
-
-import android.annotation.SuppressLint
-import android.content.Intent
-import android.content.Intent.getIntent
-import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
-import com.walkly.walkly.LoginActivity
-import com.walkly.walkly.MainActivity
-import com.walkly.walkly.R
-import kotlinx.android.synthetic.main.fragment_home.*
-
-
-@SuppressLint("Registered")
-class HomeFragment : Fragment(), View.OnClickListener {
-    lateinit var v : View
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        v =  inflater.inflate(R.layout.fragment_home, container, false)
-        return v
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-    super.onViewCreated(view, savedInstanceState)
-    signOutButton.setOnClickListener(this)
-    verifyEmailButton.setOnClickListener(this)
-    }
-
-    override fun onClick(v: View) {
-        val i = v.id
-        when (i) {
-            R.id.signOutButton -> signOut()
-        }
-    }
-
-    private fun signOut() {
-        FirebaseAuth.getInstance().signOut()
-        updateUI()
-    }
-
-
-    private fun updateUI() {
-            var intent = Intent(activity, LoginActivity::class.java)
-            startActivity(intent)
-
-    }
-
-   /* private lateinit var auth: FirebaseAuth
-    private lateinit var v : View
-    private var mAuthStateListener: FirebaseAuth.AuthStateListener? = null
-
-    // [END declare_auth]
-
-    public override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-         v = inflater.inflate(R.layout.fragment_home,container,false)
-        auth = FirebaseAuth.getInstance()
-        return v
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        emailSignInButton.setOnClickListener(this)
-        emailCreateAccountButton.setOnClickListener(this)
-        signOutButton.setOnClickListener(this)
-        verifyEmailButton.setOnClickListener(this)
-    }
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-
-        // [START initialize_auth]
-        // Initialize Firebase Auth
-        auth = FirebaseAuth.getInstance()
-        mAuthStateListener = FirebaseAuth.AuthStateListener { firebaseAuth ->
-            val firebaseUser = firebaseAuth.currentUser
-            val message: String
-            if (firebaseUser != null) {
-                message = "onAuthStateChanged signed in : " + firebaseUser.uid
-            } else {
-                message = "onAuthStateChanged signed out"
-            }
-        }
-    }
-
-    // [START on_start_check_user]
-    public override fun onStart() {
-        super.onStart()
-        // Check if user is signed in (non-null) and update UI accordingly.
-        val currentUser = auth.currentUser
-
-        mAuthStateListener?.let { auth.addAuthStateListener(it) };
-        updateUI(currentUser)
-    }
-    // [END on_start_check_user]
-
-    private fun createAccount(email: String, password: String) {
-        Log.d(TAG, "createAccount:$email")
-        if (!validateForm()) {
-            return
-        }
-
-        // [START create_user_with_email]
-        getActivity()?.let {
-            auth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(it) { task ->
-                        if (task.isSuccessful) {
-                            // Sign in success, update UI with the signed-in user's information
-                            Log.d(TAG, "createUserWithEmail:success")
-                            val user = auth.currentUser
-                            updateUI(user)
-                        } else {
-                            // If sign in fails, display a message to the user.
-                            Log.w(TAG, "createUserWithEmail:failure", task.exception)
-                            Toast.makeText(
-                                context, "Authentication failed.",
-                                Toast.LENGTH_SHORT
-                            ).show()
-                            updateUI(null)
-                        }
-
-
-                    // [START_EXCLUDE]
-                    // [END_EXCLUDE]
-                }
-        }
-        // [END create_user_with_email]
-    }
-
-    private fun signIn(email: String, password: String) {
-        Log.d(TAG, "signIn:$email")
-        if (!validateForm()) {
-            return
-        }
-
-        // [START sign_in_with_email]
-        getActivity()?.let {
-            auth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(it) { task ->
-                    if (task.isSuccessful) {
-                        // Sign in success, update UI with the signed-in user's information
-                        Log.d(TAG, "signInWithEmail:success")
-                        val user = auth.currentUser
-                        updateUI(user)
-                    } else {
-                        // If sign in fails, display a message to the user.
-                        Log.w(TAG, "signInWithEmail:failure", task.exception)
-                        Toast.makeText(context, "Authentication failed.",
-                            Toast.LENGTH_SHORT).show()
-                        updateUI(null)
-                    }
-
-                    // [START_EXCLUDE]
-                    if (!task.isSuccessful) {
-                        status.setText(R.string.auth_failed)
-                    }
-                    // [END_EXCLUDE]
-                }
-        }
-        // [END sign_in_with_email]
-    }
-
-    private fun signOut() {
-        auth.signOut()
-        updateUI(null)
-    }
-
-    private fun sendEmailVerification() {
-        // Disable button
-        verifyEmailButton.isEnabled = false
-
-        // Send verification email
-        // [START send_email_verification]
-        val user = auth.currentUser
-        getActivity()?.let {
-            user?.sendEmailVerification()
-                ?.addOnCompleteListener(it) { task ->
-                    // [START_EXCLUDE]
-                    // Re-enable button
-                    verifyEmailButton.isEnabled = true
-
-                    if (task.isSuccessful) {
-                        Toast.makeText(context,
-                            "Verification email sent to ${user.email} ",
-                            Toast.LENGTH_SHORT).show()
-                    } else {
-                        Log.e(TAG, "sendEmailVerification", task.exception)
-                        Toast.makeText(context,
-                            "Failed to send verification email.",
-                            Toast.LENGTH_SHORT).show()
-                    }
-                    // [END_EXCLUDE]
-                }
-        }
-        // [END send_email_verification]
-    }
-
-    private fun validateForm(): Boolean {
-        var valid = true
-
-        val email = fieldEmail.text.toString()
-        if (TextUtils.isEmpty(email)) {
-            fieldEmail.error = "Required."
-            valid = false
-        } else {
-            fieldEmail.error = null
-        }
-
-        val password = fieldPassword.text.toString()
-        if (TextUtils.isEmpty(password)) {
-            fieldPassword.error = "Required."
-            valid = false
-        } else {
-            fieldPassword.error = null
-        }
-
-        return valid
-    }
-
-    private fun updateUI(user: FirebaseUser?) {
-        if (user != null) {
-            status.text = getString(R.string.emailpassword_status_fmt,
-                user.email, user.isEmailVerified)
-            detail.text = getString(R.string.firebase_status_fmt, user.uid)
-
-            emailPasswordButtons.visibility = View.GONE
-            emailPasswordFields.visibility = View.GONE
-            signedInButtons.visibility = View.VISIBLE
-
-            verifyEmailButton.isEnabled = !user.isEmailVerified
-        } else {
-            status.setText(R.string.signed_out)
-            detail.text = null
-
-            emailPasswordButtons.visibility = View.VISIBLE
-            emailPasswordFields.visibility = View.VISIBLE
-            signedInButtons.visibility = View.GONE
-        }
-    }
-
-    override fun onClick(v: View) {
-        val i = v.id
-        when (i) {
-            R.id.emailCreateAccountButton -> createAccount(fieldEmail.text.toString(), fieldPassword.text.toString())
-            R.id.emailSignInButton -> signIn(fieldEmail.text.toString(), fieldPassword.text.toString())
-            R.id.signOutButton -> signOut()
-            R.id.verifyEmailButton -> sendEmailVerification()
-        }
-    }
-
-    companion object {
-        private const val TAG = "EmailPassword"
-    }*/
-}
