+++ app/src/main/res/drawable/inverted_gradient.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <gradient
+        android:endColor="#ce6b47"
+        android:startColor="#c94554"
+        android:angle="90" />
+</shape>
