+++ app/src/main/res/layout/activity_main.xml
@@ -25,6 +25,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
+        app:itemBackground="@drawable/inverted_gradient"
         style="@style/Widget.MaterialComponents.BottomNavigationView.Colored"
         app:menu="@menu/bottom_nav_menu" />
