--- app/src/main/res/layout/fragment_home.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-   <TextView
-        android:id="@+id/welcomeString"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="128dp"
-        android:text="@string/welcome_String"
-        android:textColor="@color/common_google_signin_btn_text_dark_focused"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-   <LinearLayout
-       android:id="@+id/signedInButtons"
-       android:layout_width="match_parent"
-       android:layout_height="wrap_content"
-       android:layout_centerInParent="true"
-       android:orientation="horizontal"
-       android:paddingLeft="16dp"
-       android:paddingRight="16dp"
-       android:layout_marginBottom="64dp"
-       android:weightSum="2.0"
-       app:layout_constraintBottom_toBottomOf="parent">
-
-      <Button
-          android:id="@+id/signOutButton"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_marginStart="@dimen/button_horizontal_margin"
-          android:layout_marginEnd="@dimen/button_horizontal_margin"
-          android:layout_weight="1.0"
-          android:text="@string/sign_out"
-          android:theme="@style/ThemeOverlay.MyDarkButton" />
-
-      <Button
-          android:id="@+id/verifyEmailButton"
-          android:layout_width="0dp"
-          android:layout_height="wrap_content"
-          android:layout_marginStart="@dimen/button_horizontal_margin"
-          android:layout_marginEnd="@dimen/button_horizontal_margin"
-          android:layout_weight="1.0"
-          android:text="@string/verify_email"
-          android:theme="@style/ThemeOverlay.MyDarkButton" />
-
-   </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
