+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,24 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home"
-        android:checked="true"
-        android:color="@color/design_default_color_primary_dark"/>
+
 
     <item
         android:id="@+id/navigation_dashboard"
         android:icon="@drawable/ic_dashboard_black_24dp"
-        android:title="@string/title_dashboard"
+        android:title="@string/title_battles"
         android:state_checked="false"
         android:color="@color/cardview_light_background"/>
 
     <item
+        android:id="@+id/navigation_home"
+        android:icon="@drawable/ic_home_black_24dp"
+        android:title="@string/title_map"
+        android:checked="true"
+        android:color="@color/design_default_color_primary_dark"/>
+
+    <item
         android:id="@+id/navigation_notifications"
-        android:icon="@drawable/ic_notifications_black_24dp"
-        android:title="@string/title_notifications"
+        android:icon="@drawable/ic_profile"
+        android:title="@string/title_profile"
         android:state_checked="false"
         android:color="@color/cardview_light_background"
         />
